+++ apps/ria/src/app/app.component.html
@@ -1,1 +1,1 @@
-<cvi-ng-process-diagram [boxes]="boxes"></cvi-ng-process-diagram>
+<cvi-ng-hierarchical-box-diagram [boxes]="boxes"></cvi-ng-hierarchical-box-diagram>
