+++ apps/ria/src/app/app.module.ts
@@ -9,14 +9,14 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CommonModule } from '@angular/common';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { ProcessGraphComponent } from './process-diagram/process-graph.component';
+import { HierarchicalBoxDiagramComponent } from './hierarchical-box-diagram/hierarchical-box-diagram.component';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
 }
 
 @NgModule({
-  declarations: [AppComponent, ProcessGraphComponent],
+  declarations: [AppComponent, HierarchicalBoxDiagramComponent],
   imports: [
     BrowserModule,
     UiModule,
