+++ apps/ria/src/app/hierarchical-box-diagram/box.ts
