+++ apps/ria/src/app/hierarchical-box-diagram/hierarchical-box-diagram.component.scss
