+++ apps/ria/src/app/hierarchical-box-diagram/hierarchical-box-diagram.component.spec.ts
@@ -1,17 +1,17 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ProcessGraphComponent } from './process-graph.component';
+import { HierarchicalBoxDiagramComponent } from './hierarchical-box-diagram.component';
 
 describe('ProcessDiagramComponent', () => {
-  let component: ProcessGraphComponent;
-  let fixture: ComponentFixture<ProcessGraphComponent>;
+  let component: HierarchicalBoxDiagramComponent;
+  let fixture: ComponentFixture<HierarchicalBoxDiagramComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ProcessGraphComponent],
+      declarations: [HierarchicalBoxDiagramComponent],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(ProcessGraphComponent);
+    fixture = TestBed.createComponent(HierarchicalBoxDiagramComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
